remove utils folder, fix mypy
